- Fixed a bunch of compiler warnings - Added 'Apply Changes' and 'Reset' buttons to the map object properties panel - all properties can now be changed with this except for line/thing flags in non-udmf formats - Changed the default pan view key from ctrl/command+space to alt+space, since command+space on mac seems not to be registering correctly - Map canvas no longer takes focus when the mouse cursor enters it (was taking focus away even from modal dialogs)
git-svn-id: http://mancubus.net/svn/hosted/slade/branches/map_editor@1155 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/MapEditor.cpp
@@ -733,7 +733,7 @@ unsigned MapEditor::numEditorMessages() {
 
 string MapEditor::getEditorMessage(int index) {
 	// Check index
-	if (index < 0 || index >= editor_messages.size())
+	if (index < 0 || index >= (int)editor_messages.size())
 		return "";
 
 	return editor_messages[index].message;
@@ -741,7 +741,7 @@ string MapEditor::getEditorMessage(int index) {
 
 long MapEditor::getEditorMessageTime(int index) {
 	// Check index
-	if (index < 0 || index >= editor_messages.size())
+	if (index < 0 || index >= (int)editor_messages.size())
 		return -1;
 
 	return theApp->runTimer() - editor_messages[index].act_time;
